apply some changes from the review
still gotta implement the channels
